The game is now in a semi-working state
Changed structure of Atom class.
Changed the way separation of atoms work. This enables more control of what should be separated.
Slight tweak to the way objects are selected and deselected.

The game is now at a point where a report can be started upon. There still a lot to do, but the fundamentals are there.